Moving testing module to corpuslib
