+++ corpuslib/__init__.py
@@ -1,0 +1,1 @@
+from test import Accuracy, Test
